AV1708: Do not use 'Helper' or 'Shared' in type names
