+++ test/UnitTests/Internal/TypeConversionTests.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Collections.Generic;
 using JsonApiDotNetCore;
-using JsonApiDotNetCore.Resources;
 using Xunit;
 
 namespace UnitTests.Internal
 {
-    public sealed class TypeHelperTests
+    public sealed class TypeConversionTests
     {
         [Fact]
         public void Can_Convert_DateTimeOffsets()
@@ -131,46 +130,6 @@ namespace UnitTests.Internal
             Assert.Throws<FormatException>(action);
         }
 
-        [Fact]
-        public void New_Creates_An_Instance_If_T_Implements_Interface()
-        {
-            // Arrange
-            Type type = typeof(Model);
-
-            // Act
-            var instance = (IIdentifiable)TypeHelper.CreateInstance(type);
-
-            // Assert
-            Assert.NotNull(instance);
-            Assert.IsType<Model>(instance);
-        }
-
-        [Fact]
-        public void Implements_Returns_True_If_Type_Implements_Interface()
-        {
-            // Arrange
-            Type type = typeof(Model);
-
-            // Act
-            bool result = TypeHelper.IsOrImplementsInterface(type, typeof(IIdentifiable));
-
-            // Assert
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void Implements_Returns_False_If_Type_DoesNot_Implement_Interface()
-        {
-            // Arrange
-            Type type = typeof(string);
-
-            // Act
-            bool result = TypeHelper.IsOrImplementsInterface(type, typeof(IIdentifiable));
-
-            // Assert
-            Assert.False(result);
-        }
-
         private enum TestEnum
         {
             Test = 1
@@ -186,12 +145,6 @@ namespace UnitTests.Internal
 
         private interface IType
         {
-        }
-
-        private sealed class Model : IIdentifiable
-        {
-            public string StringId { get; set; }
-            public string LocalId { get; set; }
         }
     }
 }
